% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_wordkrill.R
\name{bootstrap_wordkrill}
\alias{bootstrap_wordkrill}
\title{Compute Bootstrap Standard Errors}
\usage{
bootstrap_wordkrill(model, B = 500, K = NULL, verbose = TRUE)
}
\arguments{
\item{model}{a fitted Wordkrill model}

\item{B}{how many replications}

\item{K}{dimension of the analysed Wordkrill model}

\item{verbose}{give progress updates}
}
\value{
list with standard errors for document positions
}
\description{
Computes bootstrap standard errors for document positions from a fitted
Wordkrill model.
}
\details{
This function computes a parametric bootstrap by resampling counts from the
fitted feature counts, refitting the model, and storing the document
positions.
The standard deviations for each resampled document position are returned.
}
\author{
Benjamin Riesch
}
