% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_epsilon.R
\name{compute_epsilon}
\alias{compute_epsilon}
\title{Calculates epsilon so that the estimated target variables (mean, variance
and covariance) lie within an appropriate range around the target values.}
\usage{
compute_epsilon(dtm, alpha = 0.05, norm = T, K = 2)
}
\arguments{
\item{dtm}{a quanteda DTM object}

\item{alpha}{control parameters to determine the strictness of the epsilon
environment}

\item{norm}{logical value indicating normality distribution of the thetas is
assumed or not}

\item{K}{dimension parameter for the model to be estimated}
}
\value{
An object of class numeric, containing:

\item{epsilon}{estimated optimal epsilon for the model to be estimated}
}
\description{
Calculates epsilon so that the estimated target variables (mean, variance
and covariance) lie within an appropriate range around the target values.
}
\references{
Riesch (2025) 'Wordkrill: Extending Wordfish into the
multidimensional political space'. https://doi.org/10.48550/arXiv.2506.20275
}
\author{
Benjamin Riesch
}
