% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wordkrill.R
\name{get_document_para}
\alias{get_document_para}
\title{Get document positions}
\usage{
get_document_para(model)
}
\arguments{
\item{model}{a fitted Wordkrill model}
}
\description{
Get document positions
}
