% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_urkrill.R
\name{initialize_urkrill}
\alias{initialize_urkrill}
\title{Get initial values for a Wordkrill model with K dimensions.}
\usage{
initialize_urkrill(dtm, K, ort = F)
}
\arguments{
\item{dtm}{a document by word matrix of counts}

\item{K}{number of dimensions of the corresponding wordkrill model}

\item{ort}{logical value indicating whether the initial values are
orthogonalised}
}
\value{
List with elements:
\item{alpha}{starting values of alpha parameters}
\item{psi}{starting values of psi parameters}
\item{theta_1...theta_K}{starting values of theta_1 to theta_K}
\item{theta_1...theta_K}{starting values of theta_1 to theta_K}
\item{beta_1...beta_K}{starting values of beta_1 to beta_K}
}
\description{
This function is only called by model fitting routines and does therefore
not take a dtm classes. dtm is assumed to be in document by feature form.
}
\details{
In the poisson form of the model incidental parameters (alpha) are set to
logged rowmeans divided by the first entry of rowmeans. Intercept (psi)
values are set to log(colmeans). These are subtracted from a the data matrix,
which is logged and decomposed by SVD. Word slope (beta) and document
position (theta) are estimated by rescaling SVD output.
}
\references{
Slapin and Proksch (2008) 'A Scaling Model for Estimating
Time-Series Party Positions from Texts.' American Journal of Political
Science 52(3):705-772.

Lowe (2020) 'Austin: Do Things with Words'.
https://conjugateprior.github.io/austin

Riesch (2025) 'Wordkrill: Extending Wordfish into the
multidimensional political space'. https://doi.org/10.48550/arXiv.2506.20275
}
\author{
Benjamin Riesch

This is a multidimensional extension of the algorithm proposed by
Slapin and Proksch for obtaining initial values.
}
