% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wordkrill.R
\name{print.summary.wordkrill}
\alias{print.summary.wordkrill}
\title{Print summarised Wordkrill model}
\usage{
\method{print}{summary.wordkrill}(x, ...)
}
\arguments{
\item{x}{a summary.wordkrill object}

\item{...}{additional arguments}
}
\description{
Print summarised Wordkrill model
}
