% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_wordkrill.R
\name{summary.wordkrill}
\alias{summary.wordkrill}
\title{Summarizes a Wordkrill model}
\usage{
\method{summary}{wordkrill}(object, ...)
}
\arguments{
\item{object}{a fitted Wordkrill model}

\item{...}{additional arguments (e.g., top_n to select number of displayed
features)}
}
\description{
Summarizes a Wordkrill model
}
